reverting changes cause git is stupid
