+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/PlayerAlarm/AlarmOverlay.java
@@ -1,4 +1,4 @@
- package net.runelite.client.plugins.microbot.storm.plugins.PlayerAlarm;
+ package net.runelite.client.plugins.microbot.PlayerAlarm;
 
  import java.awt.Color;
  import java.awt.Dimension;
