+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/blackjack/BlackJackConfig.java
@@ -1,10 +1,11 @@
-package net.runelite.client.plugins.microbot.storm.plugins.blackjack;
+package net.runelite.client.plugins.microbot.blackjack;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 import net.runelite.client.config.ConfigSection;
-import net.runelite.client.plugins.microbot.storm.plugins.blackjack.enums.Thugs;
+import net.runelite.client.plugins.microbot.blackjack.enums.Thugs;
+import net.runelite.client.plugins.microbot.util.misc.Rs2Food;
 
 @ConfigGroup("BlackJack")
 public interface BlackJackConfig extends Config {
