+++ runelite-client/src/main/java/net/runelite/client/plugins/microbot/blackjack/BlackJackOverlay.java
@@ -1,4 +1,5 @@
-package net.runelite.client.plugins.microbot.storm.plugins.BankStander;
+package net.runelite.client.plugins.microbot.blackjack;
+
 
 import net.runelite.client.plugins.microbot.Microbot;
 import net.runelite.client.ui.overlay.OverlayPanel;
@@ -9,26 +10,23 @@ import net.runelite.client.ui.overlay.components.TitleComponent;
 import javax.inject.Inject;
 import java.awt.*;
 
-public class BankStanderOverlay extends OverlayPanel {
+public class BlackJackOverlay extends OverlayPanel {
     @Inject
-    BankStanderOverlay(BankStanderPlugin plugin) {
+    BlackJackOverlay(BlackJackPlugin plugin)
+    {
         super(plugin);
         setPosition(OverlayPosition.TOP_LEFT);
         setNaughty();
     }
-
     @Override
     public Dimension render(Graphics2D graphics) {
         try {
             panelComponent.setPreferredSize(new Dimension(200, 300));
             panelComponent.getChildren().add(TitleComponent.builder()
-                    .text("Storm's BankStander V" + BankStanderScript.version)
-                    .color(BankStanderPlugin.fontColor)
+                    .text("Blackjack thieving training : " + BlackJackScript.version)
+                    .color(Color.GREEN)
                     .build());
-            panelComponent.getChildren().add(LineComponent.builder()
-                    .left("Items processed : " + BankStanderScript.itemsProcessed)
-                    .leftColor(BankStanderPlugin.fontColor)
-                    .build());
+
             panelComponent.getChildren().add(LineComponent.builder().build());
 
             panelComponent.getChildren().add(LineComponent.builder()
@@ -36,7 +34,7 @@ public class BankStanderOverlay extends OverlayPanel {
                     .build());
 
 
-        } catch (Exception ex) {
+        } catch(Exception ex) {
             System.out.println(ex.getMessage());
         }
         return super.render(graphics);
